+++ test/data/1024.PS.EDS/1024.p_s
@@ -1,0 +1,6 @@
+..2
+1024(1).psref
+1
+#1 SPOT 1
+1024(1)_pt1.psmsa
+360.0,253.0,731.0,547.0
