+++ test/data/128.PS.EDS/128.p_s
@@ -1,0 +1,6 @@
+..2
+128(1).psref
+1
+#1 SPOT 1
+128(1)_pt1.psmsa
+385.0,263.0,731.0,547.0
