+++ test/data/2048.PS.EDS/2048.p_s
@@ -1,0 +1,6 @@
+..2
+2048(1).psref
+1
+#1 SPOT 1
+2048(1)_pt1.psmsa
+378.0,246.0,731.0,547.0
