+++ test/data/256.PS.EDS/256.p_s
@@ -1,0 +1,6 @@
+..2
+256(1).psref
+1
+#1 SPOT 1
+256(1)_pt1.psmsa
+368.0,269.0,731.0,547.0
