+++ test/data/512.PS.EDS/512.p_s
@@ -1,0 +1,6 @@
+..2
+512(1).psref
+1
+#1 SPOT 1
+512(1)_pt1.psmsa
+396.0,265.0,731.0,547.0
