+++ test/data/64.PS.EDS/64.p_s
@@ -1,0 +1,6 @@
+..2
+64(1).psref
+1
+#1 SPOT 1
+64(1)_pt1.psmsa
+394.0,241.0,731.0,547.0
