+++ test/data/jeol-images/3.txt
@@ -1,0 +1,94 @@
+$SEM_DATA_VERSION 1
+$CM_SIGNAL_NAME COMPO
+$CM_DETECTOR_NAME COMPO
+$$SM_ADD_IMAGE 
+$$SM_ADD_CHECKED_CH 1
+$$ADD_CH1_SIGNAL SEI
+$$ADD_CH2_SIGNAL SEI
+$$ADD_CH3_SIGNAL SEI
+$$ADD_CH4_SIGNAL SEI
+$$ADD_CH1_DETECTOR SEI
+$$ADD_CH2_DETECTOR SEI
+$$ADD_CH3_DETECTOR SEI
+$$ADD_CH4_DETECTOR SEI
+$CM_CONTRAST 3463
+$CM_BRIGHTNESS 112
+$$SM_MIX_IMAGE 
+$CM_ACCEL_VOLT 15.00
+$SM_ARRIVAL_EMI 0.00
+$SM_GB_GUN_VOLT 15.00
+$SM_GB_BIAS_VOLT 0.00
+$$SM_WD 10.73
+$CM_MAG 300
+$$SM_COLUMN_MODE NOR
+$$SM_OL_APERTURE 2
+$$SM_CL_COARSE 50
+$$SM_CL_FINE 140
+$$SM_GUN_ALI_TILT 9 -19
+$$SM_CL_STIG 0 0
+$$SM_OL_STIG -80 -23
+$$SM_OL_STIG_CENTER 0 0 0 0
+$$SM_OLAP_ALI 0 0
+$$SM_PROBE_NO 51
+$$SM_PROBE_FINE 371
+$$SM_ACL_ALI 0 0
+$$SM_COLUMN_CENTER 0 0
+$$SM_EF_ALI 0 0
+$$SM_SCAN_ROTATION 0.00
+$$SM_WD_CORRECT 0
+$$SM_ACL_CORRECT 0
+$$SM_INTEG_NO 1
+$$CM_SCAN_MODE 4
+$$CM_SCAN_SPEED 9
+$$CM_INSTRUMENT_SCAN_SPEED 9
+$$SM_IMAGE_INTEGRATION_TYPE 0
+$$CM_PROCESS_TIME 128.00
+$$SM_PHOTO_SN_NUMBER 1
+$$SM_R_FILTER_MODE 0
+$$SM_SEI_COLLECTOR 0
+$$SM_R_FILTER_INDEX 0
+$$SM_DET_SE1_MODE_NUMBER 0
+$$SM_DET_TTL_MODE_NUMBER 0
+$$SM_DET_SE2_MODE_NUMBER 0
+$$SM_NE_MODE_NUMBER 0
+$$SM_TTL_COLLECTOR_VOLTAGE 0.00
+$$SM_SED_FILTER_MODE 0
+$$SM_UED_FILTER_MODE 0
+$$SM_UED_FILTER_VALUE 0.00
+$$SM_ACB_MODE 1
+$$SM_ACB_CONTRAST 5.00
+$$SM_ACB_BRIGHTNESS -3.00
+$CM_SIGNAL COMPO
+$CM_DETECTOR COMPO
+$CM_STAGE_POS 16.5176 -22.9579 12.7350 0 0 0
+$SM_EMI_CURRENT 61.60
+$$SM_PROBE_CURRENT 3.66e-009
+$SM_PENNING_VAC 35.000
+$SM_PRESSURE_AT_LOW_VACUUM 0.000
+$SM_VACUUM_MODE_FOR_SPECIMEN_CHAMBER 0
+$CM_FORMAT JEOL-SEM
+$CM_VERSION 3
+$CM_COMMENT 
+$CM_DATE 2019/09/30
+$CM_TIME 10:30:36
+$CM_OPERATOR UMNUser
+$CM_INSTRUMENT 8530F
+$$SM_MICRON_BAR 64
+$$SM_MICRON_MARKER 10um
+$CM_FRAME_SIZE 0 0
+$CM_TITLE 20190930_001
+$CM_LABEL UMN
+$CM_OFFSET 0 0
+$CM_FULL_SIZE 2560 1920
+$AN_CURSOR_MEAS_TYPE 0
+$AN_EXPORT_MODE 1
+$CM_COLOR_MODE 0
+$AN_ANNOTATION_MODE 0
+$AN_CURSOR_MEAS $AN_LINE%0 $AN_LINE%1 $AN_LINE%2 $AN_LINE%3 $AN_TEXT%0 $AN_TEXT%1 $AN_TEXT%2
+$AN_LINE%0 480.00000000 0.00000000 480.00000000 959.00000000 65280 1.00000000 0 0
+$AN_LINE%1 800.00000000 0.00000000 800.00000000 959.00000000 65280 1.00000000 0 0
+$AN_LINE%2 0.00000000 360.00000000 1279.00000000 360.00000000 65280 1.00000000 0 0
+$AN_LINE%3 0.00000000 600.00000000 1279.00000000 600.00000000 65280 1.00000000 0 0
+$AN_TEXT%0 5.00000000 874.00000000 215.00000000 910.00000000 "X:0.827µm" "Courier New" 0 32.00000000 16777215 65793 1 0
+$AN_TEXT%1 220.00000000 874.00000000 430.00000000 910.00000000 "Y:13.235µm" "Courier New" 0 32.00000000 16777215 65793 1 0
+$AN_TEXT%2 435.00000000 874.00000000 645.00000000 910.00000000 "D:13.261µm" "Courier New" 0 32.00000000 16777215 65793 1 0
