--- tests/data/1024(1).PS.EDS/1024(1).p_s
@@ -1,6 +1,0 @@
-..2
-1024(1).psref
-1
-#1 SPOT 1
-1024(1)_pt1.psmsa
-360.0,253.0,731.0,547.0
