--- tests/data/4096(1).PS.EDS/4096(1).p_s
@@ -1,6 +1,0 @@
-..2
-4096(1).psref
-1
-#1 SPOT 1
-4096(1)_pt1.psmsa
-384.0,256.0,731.0,547.0
