--- tests/data/512(1).PS.EDS/512(1).p_s
@@ -1,6 +1,0 @@
-..2
-512(1).psref
-1
-#1 SPOT 1
-512(1)_pt1.psmsa
-396.0,265.0,731.0,547.0
