--- tests/data/64(1).PS.EDS/64(1).p_s
@@ -1,6 +1,0 @@
-..2
-64(1).psref
-1
-#1 SPOT 1
-64(1)_pt1.psmsa
-394.0,241.0,731.0,547.0
